+++ .eslintrc.js
@@ -13,7 +13,8 @@ module.exports = {
   ],
   root: true,
   env: {
-    node: true
+    node: true,
+    jest: true,
   },
   rules: {
     '@typescript-eslint/interface-name-prefix': 'off',